Adding casts to make printf happy.

+++ runtime/basis/Word32/mulOverflow.c
@@ -9,6 +9,6 @@ Word Word32_mulOverflow(Word lhs, Word rhs, Bool *overflow) {
 	*overflow = (tmp != (Word)tmp);
 	if (FALSE)
 		fprintf (stderr, "0x%x * 0x%x = 0x%llu (overflow = %d)\n",
-				lhs, rhs, tmp, *overflow);
+				(uint)lhs, (uint)rhs, tmp, *overflow);
 	return tmp;
 }
